# KAFKA-9670: Reduce allocations in Metadata Response preparation